velocidb::btree

Struct BTree

pub struct BTree {
    root_page: Arc<RwLock<PageId>>,
    pager: Arc<RwLock<Pager>>,
}

Fields
root_page: Arc<RwLock<PageId>>pager: Arc<RwLock<Pager>>
Implementations

impl BTree

pub fn new(pager: Arc<RwLock<Pager>>) -> Result<Self>

pub fn from_root(root_page: PageId, pager: Arc<RwLock<Pager>>) -> Self

pub fn insert(&mut self, key: i64, row: &Row) -> Result<()>

pub fn search(&self, key: i64) -> Result<Option<Row>>

pub fn scan(&self) -> Result<Vec<(i64, Row)>>

fn scan_node( &self, page_id: PageId, results: &mut Vec<(i64, Row)>, ) -> Result<()>

pub fn delete(&mut self, key: i64) -> Result<bool>

fn handle_leaf_underflow( &self, pager: &mut Pager, page_id: PageId, root_page: PageId, ) -> Result<()>

fn find_sibling_info( &self, parent_page: &Page, parent_header: &NodeHeader, child_id: PageId, ) -> Result<(PageId, i64, bool)>

fn redistribute_from_sibling( &self, pager: &mut Pager, page_id: PageId, sibling_id: PageId, parent_id: PageId, separator_key: i64, sibling_is_left: bool, ) -> Result<()>

fn read_entry_at_index( &self, data: &[u8], start_offset: usize, index: usize, ) -> Result<(i64, usize, Vec<u8>, usize)>

fn update_parent_key( &self, pager: &mut Pager, parent_id: PageId, old_key: i64, new_key: i64, ) -> Result<()>

fn merge_leaves( &self, pager: &mut Pager, page_id: PageId, sibling_id: PageId, parent_id: PageId, _separator_key: i64, root_page: PageId, sibling_is_left: bool, ) -> Result<()>

fn remove_child_from_parent( &self, pager: &mut Pager, parent_id: PageId, child_id: PageId, root_page: PageId, ) -> Result<()>

fn handle_internal_underflow( &self, pager: &mut Pager, page_id: PageId, root_page: PageId, ) -> Result<()>

fn parse_internal_node( &self, data: &[u8], header: &NodeHeader, ) -> Result<(Vec<i64>, Vec<PageId>)>

Reads an internal node’s keys and child pointers into vectors for easier manipulation during merge/redistribute.

fn write_internal_body( &self, page: &mut Page, keys: &[i64], children: &[PageId], ) -> Result<()>

Writes the body of an internal node (children + interleaved keys). The caller is responsible for writing the NodeHeader.

fn redistribute_internal( &self, pager: &mut Pager, page_id: PageId, sibling_id: PageId, parent_id: PageId, separator_key: i64, sibling_is_left: bool, ) -> Result<()>

fn merge_internal( &self, pager: &mut Pager, page_id: PageId, sibling_id: PageId, parent_id: PageId, separator_key: i64, root_page: PageId, sibling_is_left: bool, ) -> Result<()>

fn find_leaf( &self, pager: &mut Pager, root_page: PageId, key: i64, ) -> Result<PageId>

fn insert_into_leaf( &self, pager: &mut Pager, page_id: PageId, key: i64, data: &[u8], ) -> Result<Option<PageId>>

fn find_data_end(&self, header: &NodeHeader, data: &[u8]) -> Result<usize>

fn split_leaf_node( &self, pager: &mut Pager, page_id: PageId, ) -> Result<(PageId, i64)>

fn insert_into_parent( &self, pager: &mut Pager, left_page: PageId, right_page: PageId, split_key: i64, ) -> Result<Option<PageId>>

fn create_new_root( &self, pager: &mut Pager, left_page: PageId, right_page: PageId, split_key: i64, ) -> Result<PageId>

fn set_parent_pointer( &self, pager: &mut Pager, page_id: PageId, new_parent: u32, ) -> Result<()>

Reads the page, updates its parent field, and writes it back through the pager so the change is durable (logged in the WAL).

fn insert_into_internal( &self, pager: &mut Pager, page_id: PageId, left_page: PageId, right_page: PageId, key: i64, ) -> Result<Option<PageId>>

fn split_internal_node( &self, pager: &mut Pager, page_id: PageId, _left_page: PageId, right_page: PageId, key: i64, ) -> Result<Option<PageId>>

fn serialize_row(&self, row: &Row) -> Result<Vec<u8>>

fn deserialize_row(&self, data: &[u8]) -> Result<Row>

pub fn root_page(&self) -> PageId